// src/enums/App.ts
export enum Access {
  SHARED = 'shared',
  PRIVATE = 'private'
}

export enum IconType {
  HERO_ICON = 'heroicon',
  IMAGE = 'image'
}

export enum PageClass {
  PAGE = 'page',
  INTEGRATION = 'integration'
}

export enum PageType {
  OVERVIEW = 'overview',
  FILES = 'files',
  TASKS = 'tasks',
  MESSAGING = 'messaging',
  PATH_SETTINGS = 'path-settings',
  ALL_PATHS = 'all-paths',
  TRELLO_INTEG = 'trello-integ',
  SIGN_NOW_INTEG = 'signnow-integ',
  SQUARE_INTEG = 'square-integ',
  GMAIL_INTEG = 'gmail-integ',
  GDRIVE_INTEG = 'gdrive-integ',
  GITHUB_INTEG = 'github-integ',
  UPWORK_INTEG = 'upwork-integ',
  FIGMA_PROTOTYPE_EMBED = 'figma-prototype-embed',
  FIGMA_DESIGN_EMBED = 'figma-design-embed',
  IFRAME = 'iframe',
  WAVEAPPS_INTEG = 'waveapps-integ',
  TRASH = 'trash',
  THIO_CHAT_AI = 'thio-chat-ai'
}

export enum DraggableMode {
  GRID = 'grid',
  FLOAT = 'float'
}